Conflicts:
	CMakeLists.txt
	main/include/eudaq/Event.hh
	main/include/eudaq/FileWriter.hh
	main/include/eudaq/Serializer.hh

+++ main/include/eudaq/Clusters.hh
@@ -18,14 +18,14 @@ namespace eudaq {
       void read(std::istream & in);
       void SetEventNum(int eve);
       int EventNum() const { return m_event; }
-      unsigned long long EventTimestamp() const { return m_timestamp; }
+      uint64_t EventTimestamp() const { return m_timestamp; }
       int NumClusters() const { return m_x.size(); }
       int ClusterX(int i) const { return m_x[i]; }
       int ClusterY(int i) const { return m_y[i]; }
       int ClusterA(int i) const { return m_a[i]; }
     private:
       int m_event;
-      unsigned long long m_timestamp;
+      uint64_t m_timestamp;
       std::vector<int> m_x, m_y, m_a;
   };
