+++ main/include/eudaq/FileWriter.hh
@@ -14,7 +14,7 @@ namespace eudaq {
       FileWriter();
       virtual void StartRun(unsigned runnumber) = 0;
 	  virtual void WriteEvent(const DetectorEvent &) = 0;
-      virtual unsigned long long FileBytes() const = 0;
+      virtual uint64_t FileBytes() const = 0;
       void SetFilePattern(const std::string & p) { m_filepattern = p; }
       virtual ~FileWriter() {}
     protected:
