+++ main/include/eudaq/RunControl.hh
@@ -53,13 +53,13 @@ namespace eudaq {
       bool m_done;
       bool m_listening;
     protected:
-      long m_runnumber;   ///< The current run number
+      int32_t m_runnumber;   ///< The current run number
       TransportServer * m_cmdserver; ///< Transport for sending commands
 	std::unique_ptr<std::thread> m_thread;
       size_t m_idata, m_ilog;
       std::string m_logaddr; // address of log collector
       std::map<size_t,std::string> m_dataaddr; // map of data collector addresses
-      long long m_runsizelimit;
+      int64_t m_runsizelimit;
       bool m_stopping, m_busy, m_producerbusy;
   };
